# backend/docs/contracts/generated-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UserManagementController_findAll",
        "summary": "List all users",
        "description": "Returns a paginated list of users with optional filtering by status, department, role, and search term. Requires User.READ permission.",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by user status (ACTIVE, DISABLED, PENDING)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "description": "Filter by department ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20)",
            "schema": {
              "example": "20",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name, personnel code, or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleCode",
            "required": false,
            "in": "query",
            "description": "Filter by role code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - User.READ permission required"
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "UserManagementController_create",
        "summary": "Create user (admin)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UserManagementController_getMe",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UserManagementController_updateMe",
        "summary": "Update current user profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/stats": {
      "get": {
        "operationId": "UserManagementController_getStats",
        "summary": "Get user statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/personnel/{personnelCode}": {
      "get": {
        "operationId": "UserManagementController_findByPersonnelCode",
        "summary": "Get user by personnel code",
        "parameters": [
          {
            "name": "personnelCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserManagementController_findOne",
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UserManagementController_update",
        "summary": "Update user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UserManagementController_remove",
        "summary": "Delete user (soft)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/restore": {
      "post": {
        "operationId": "UserManagementController_restore",
        "summary": "Restore deleted user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/status": {
      "put": {
        "operationId": "UserManagementController_toggleStatus",
        "summary": "Toggle user status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/roles": {
      "post": {
        "operationId": "UserManagementController_assignRole",
        "summary": "Assign role to user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/roles/{roleId}": {
      "put": {
        "operationId": "UserManagementController_updateRoleAssignment",
        "summary": "Update role assignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UserManagementController_removeRole",
        "summary": "Remove role from user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/permissions": {
      "get": {
        "operationId": "UserManagementController_getPermissions",
        "summary": "Get user effective permissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["User Management"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/audit/user/{userId}": {
      "get": {
        "operationId": "AuditController_getUserActivity",
        "summary": "Get user activity",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Audit"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IRIB DWP API",
    "description": "IRIB East Azerbaijan Digital Workplace Platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile phone number",
            "example": "09123456789"
          },
          "nameFa": {
            "type": "string",
            "description": "Persian name of the user",
            "example": "جان دو"
          },
          "nationalCode": {
            "type": "string",
            "description": "National code",
            "example": "1234567890"
          },
          "departmentIds": {
            "description": "Array of department IDs",
            "example": ["dept-1", "dept-2"],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "personnelCode": {
            "type": "string",
            "description": "Personnel code for the user",
            "example": "12345"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile phone number",
            "example": "09123456789"
          },
          "nameFa": {
            "type": "string",
            "description": "Persian name of the user",
            "example": "جان دو"
          },
          "nationalCode": {
            "type": "string",
            "description": "National code",
            "example": "1234567890"
          },
          "departmentIds": {
            "description": "Array of department IDs",
            "example": ["dept-1", "dept-2"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "initialStatus": {
            "type": "string",
            "description": "Initial status of the user",
            "example": "ACTIVE"
          }
        },
        "required": ["personnelCode", "name"]
      },
...
    }
  }
}